requirements.txt
Django==4.0.2
djangorestframework==3.13.1
numpy==1.19.2
pandas==1.2.4
torch==2.0.0
gym==0.21.0
gym-notices==0.0.8
gunicorn==20.1.0
stable-baselines3==1.6.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
redis==4.6.0
django-redis==5.3.0